+++ src/main/java/de/elomagic/spps/jshiro/SimpleCryptException.java
@@ -1,0 +1,32 @@
+/*
+ * Simple Password Protection Solution with Apache Shiro
+ *
+ * Copyright © 2021-present Carsten Rambow (spps.dev@elomagic.de)
+ *
+ * This file is part of Simple Password Protection Solution with Apache Shiro.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.elomagic.spps.jshiro;
+
+public class SimpleCryptException extends RuntimeException {
+
+    public SimpleCryptException(String message) {
+        super(message);
+    }
+
+    public SimpleCryptException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}